fix: extract keystore creation from deposit hook
